separating concerns, built controller files for queries
